WIP: add origin create resource
I've tried to map the components required by the API, and I'm still
trying to understand how the provider works. In it's current state, it
fails, but I'm not too sure why. Something to look at post holiday! I
need to add additional debugging.

I implemented a small request function to use for making requests to
edgecast in a generic way, and implemented types which eventually can be
used to work with the requests and response json data.

The creation resource should create a new resource, and set the OriginID
as the ID of the resource so it can be referred to when making requests
for other resources.